# Conflicts:
#	features/merchant/product_detail/src/main/java/com/tokopedia/product/detail/data/util/DynamicProductDetailMapper.kt
#	features/merchant/product_share/src/main/java/com/tokopedia/product/share/ProductShare.kt
#	libraries/universal_sharing/src/main/java/com/tokopedia/universal_sharing/view/bottomsheet/UniversalShareBottomSheet.kt

+++ features/content/kol_common/src/main/java/com/tokopedia/kolcommon/domain/usecase/LikeKolPostUseCase.kt
@@ -22,9 +22,9 @@ class LikeKolPostUseCase @Inject constructor(
 
         private const val LIKE_SUCCESS = 1
 
-        fun getParam(postId: Int, action: LikeKolPostAction): RequestParams {
+        fun getParam(postId: Long, action: LikeKolPostAction): RequestParams {
             return RequestParams.create().apply {
-                putInt(PARAM_ID_POST, postId)
+                putLong(PARAM_ID_POST, postId)
                 putInt(PARAM_ACTION, action.actionValue)
             }
         }
